Remove API provider abstraction complexity
- Removed APIManager, BaseAPIProvider, MockAPIProvider, ReplicateAPIProvider, WebSimAPIProvider classes (~600 lines)
- RoomManager now calls ReplicateAPI service directly for image generation
- InteractionController uses simple pattern matching for command parsing instead of API-based parsing
- Updated MemoryPalaceCore to remove API provider initialization and dependencies
- Changed configuration from apiProvider: 'mock' to enableImageGeneration: true
- Eliminates mock provider accidentally being used in production
- Simplifies architecture by removing unnecessary abstraction layer

Addresses issue #73 - API provider complexity causing mock image generation when it shouldn't
